+++ tests/wallet/test_coin_selection.py
@@ -81,7 +81,7 @@ class TestCoinSelection:
         for target_amount in coin_amounts[:100]:  # select the first 100 values
             result: Set[Coin] = await select_coins(
                 spendable_amount,
-                DEFAULT_CONSTANTS.MAX_COIN_AMOUNT,
+                uint64(DEFAULT_CONSTANTS.MAX_COIN_AMOUNT),
                 coin_list,
                 {},
                 logging.getLogger("test"),
@@ -110,7 +110,7 @@ class TestCoinSelection:
             print("Target amount: ", target_amount)
             result: Set[Coin] = await select_coins(
                 spendable_amount,
-                DEFAULT_CONSTANTS.MAX_COIN_AMOUNT,
+                uint64(DEFAULT_CONSTANTS.MAX_COIN_AMOUNT),
                 coin_list,
                 {},
                 logging.getLogger("test"),
@@ -130,7 +130,7 @@ class TestCoinSelection:
         for target_amount in [50000, 25000, 15000, 10000, 9000, 3000]:  # select the first 100 values
             dusty_result: Set[Coin] = await select_coins(
                 spendable_amount,
-                DEFAULT_CONSTANTS.MAX_COIN_AMOUNT,
+                uint64(DEFAULT_CONSTANTS.MAX_COIN_AMOUNT),
                 coin_list,
                 {},
                 logging.getLogger("test"),
@@ -161,7 +161,7 @@ class TestCoinSelection:
         for target_amount in [20000, 15000, 10000, 5000]:  # select the first 100 values
             dusty_below_target: Set[Coin] = await select_coins(
                 spendable_amount,
-                DEFAULT_CONSTANTS.MAX_COIN_AMOUNT,
+                uint64(DEFAULT_CONSTANTS.MAX_COIN_AMOUNT),
                 new_coin_list,
                 {},
                 logging.getLogger("test"),
@@ -192,7 +192,7 @@ class TestCoinSelection:
         for target_amount in [50000, 10001, 10000, 9999]:
             dusty_below_target: Set[Coin] = await select_coins(
                 spendable_amount,
-                DEFAULT_CONSTANTS.MAX_COIN_AMOUNT,
+                uint64(DEFAULT_CONSTANTS.MAX_COIN_AMOUNT),
                 new_coin_list,
                 {},
                 logging.getLogger("test"),
@@ -218,7 +218,7 @@ class TestCoinSelection:
             for target_amount in [10000, 9999]:
                 await select_coins(
                     spendable_amount,
-                    DEFAULT_CONSTANTS.MAX_COIN_AMOUNT,
+                    uint64(DEFAULT_CONSTANTS.MAX_COIN_AMOUNT),
                     coin_list,
                     {},
                     logging.getLogger("test"),
@@ -229,7 +229,7 @@ class TestCoinSelection:
             for target_amount in [10001, 20000]:
                 await select_coins(
                     spendable_amount,
-                    DEFAULT_CONSTANTS.MAX_COIN_AMOUNT,
+                    uint64(DEFAULT_CONSTANTS.MAX_COIN_AMOUNT),
                     coin_list,
                     {},
                     logging.getLogger("test"),
@@ -249,7 +249,7 @@ class TestCoinSelection:
         target_amount = uint128(40)
         exact_match_result: Set[Coin] = await select_coins(
             spendable_amount,
-            DEFAULT_CONSTANTS.MAX_COIN_AMOUNT,
+            uint64(DEFAULT_CONSTANTS.MAX_COIN_AMOUNT),
             coin_list,
             {},
             logging.getLogger("test"),
@@ -263,7 +263,7 @@ class TestCoinSelection:
         target_amount = uint128(153)
         match_2: Set[Coin] = await select_coins(
             spendable_amount,
-            DEFAULT_CONSTANTS.MAX_COIN_AMOUNT,
+            uint64(DEFAULT_CONSTANTS.MAX_COIN_AMOUNT),
             coin_list,
             {},
             logging.getLogger("test"),
@@ -276,7 +276,7 @@ class TestCoinSelection:
         target_amount = uint128(541)
         match_3: Set[Coin] = await select_coins(
             spendable_amount,
-            DEFAULT_CONSTANTS.MAX_COIN_AMOUNT,
+            uint64(DEFAULT_CONSTANTS.MAX_COIN_AMOUNT),
             coin_list,
             {},
             logging.getLogger("test"),
@@ -290,7 +290,7 @@ class TestCoinSelection:
         target_amount = spendable_amount
         match_all: Set[Coin] = await select_coins(
             spendable_amount,
-            DEFAULT_CONSTANTS.MAX_COIN_AMOUNT,
+            uint64(DEFAULT_CONSTANTS.MAX_COIN_AMOUNT),
             coin_list,
             {},
             logging.getLogger("test"),
@@ -310,7 +310,7 @@ class TestCoinSelection:
         target_amount = uint128(625)
         smallest_result: Set[Coin] = await select_coins(
             greater_spendable_amount,
-            DEFAULT_CONSTANTS.MAX_COIN_AMOUNT,
+            uint64(DEFAULT_CONSTANTS.MAX_COIN_AMOUNT),
             greater_coin_list,
             {},
             logging.getLogger("test"),
@@ -328,7 +328,7 @@ class TestCoinSelection:
         target_amount = uint128(50000)
         single_greater_result: Set[Coin] = await select_coins(
             single_greater_spendable_amount,
-            DEFAULT_CONSTANTS.MAX_COIN_AMOUNT,
+            uint64(DEFAULT_CONSTANTS.MAX_COIN_AMOUNT),
             single_greater_coin_list,
             {},
             logging.getLogger("test"),
@@ -348,7 +348,7 @@ class TestCoinSelection:
         target_amount = uint128(70000)
         multiple_greater_result: Set[Coin] = await select_coins(
             multiple_greater_spendable_amount,
-            DEFAULT_CONSTANTS.MAX_COIN_AMOUNT,
+            uint64(DEFAULT_CONSTANTS.MAX_COIN_AMOUNT),
             multiple_greater_coin_list,
             {},
             logging.getLogger("test"),
@@ -378,7 +378,7 @@ class TestCoinSelection:
         target_amount = spendable_amount - 1
         result: Set[Coin] = await select_coins(
             spendable_amount,
-            DEFAULT_CONSTANTS.MAX_COIN_AMOUNT,
+            uint64(DEFAULT_CONSTANTS.MAX_COIN_AMOUNT),
             coin_list,
             {},
             logging.getLogger("test"),
@@ -444,7 +444,7 @@ class TestCoinSelection:
             for min_coin_amount in [10, 100, 200, 300, 1000]:
                 result: Set[Coin] = await select_coins(
                     spendable_amount,
-                    DEFAULT_CONSTANTS.MAX_COIN_AMOUNT,
+                    uint64(DEFAULT_CONSTANTS.MAX_COIN_AMOUNT),
                     coin_list,
                     {},
                     logging.getLogger("test"),
@@ -477,7 +477,7 @@ class TestCoinSelection:
         # test that excluded coins are not included in the result
         selected_coins: Set[Coin] = await select_coins(
             spendable_amount,
-            DEFAULT_CONSTANTS.MAX_COIN_AMOUNT,
+            uint64(DEFAULT_CONSTANTS.MAX_COIN_AMOUNT),
             spendable_wallet_coin_records,
             {},
             logging.getLogger("test"),
@@ -495,7 +495,7 @@ class TestCoinSelection:
         with pytest.raises(ValueError):
             await select_coins(
                 spendable_amount,
-                DEFAULT_CONSTANTS.MAX_COIN_AMOUNT,
+                uint64(DEFAULT_CONSTANTS.MAX_COIN_AMOUNT),
                 spendable_wallet_coin_records,
                 {},
                 logging.getLogger("test"),
@@ -516,7 +516,7 @@ class TestCoinSelection:
         target_amount = uint128(0)
         zero_amount_result: Set[Coin] = await select_coins(
             spendable_amount,
-            DEFAULT_CONSTANTS.MAX_COIN_AMOUNT,
+            uint64(DEFAULT_CONSTANTS.MAX_COIN_AMOUNT),
             coin_list,
             {},
             logging.getLogger("test"),
